[REF] Parser: move parsing functions from nodes to plugins
