+++ packages/core/src/VNodes/VNode.ts
@@ -67,7 +67,6 @@ export class VNode {
     // Lifecycle
     //--------------------------------------------------------------------------
 
-    static parse: (node: Node) => VNode[];
     /**
      * Render the VNode to the given format.
      *
