+++ packages/plugin-heading/HeadingNode.ts
@@ -9,9 +9,4 @@ export class HeadingNode extends VElement {
     get name(): string {
         return super.name + ': ' + this.level;
     }
-    static parse(node: Node): HeadingNode[] {
-        if (['H1', 'H2', 'H3', 'H4', 'H5', 'H6'].includes(node.nodeName)) {
-            return [new HeadingNode(parseInt(node.nodeName[1]))];
-        }
-    }
 }
