+++ packages/plugin-linebreak/LineBreak.ts
@@ -3,5 +3,10 @@ import { LineBreakNode } from './LineBreakNode';
 import { ParsingFunction } from '../core/src/Parser';
 
 export class LineBreak extends JWPlugin {
-    static readonly parsingFunctions: Array<ParsingFunction> = [LineBreakNode.parse];
+    static readonly parsingFunctions: Array<ParsingFunction> = [LineBreak.parse];
+    static parse(node: Node): LineBreakNode[] {
+        if (node.nodeName === 'BR') {
+            return [new LineBreakNode()];
+        }
+    }
 }
