+++ packages/plugin-paragraph/Paragraph.ts
@@ -3,5 +3,10 @@ import { ParsingFunction } from '../core/src/Parser';
 import { ParagraphNode } from './ParagraphNode';
 
 export class Paragraph extends JWPlugin {
-    static readonly parsingFunctions: Array<ParsingFunction> = [ParagraphNode.parse];
+    static readonly parsingFunctions: Array<ParsingFunction> = [Paragraph.parse];
+    static parse(node: Node): ParagraphNode[] {
+        if (node.nodeName === 'P') {
+            return [new ParagraphNode()];
+        }
+    }
 }
