+++ packages/plugin-paragraph/ParagraphNode.ts
@@ -4,9 +4,4 @@ export class ParagraphNode extends VElement {
     constructor() {
         super('P');
     }
-    static parse(node: Node): ParagraphNode[] {
-        if (node.nodeName === 'P') {
-            return [new ParagraphNode()];
-        }
-    }
 }
